mkdir -p /etc/portage/package.keywords
Gentoo stage 3 does not currently provide /etc/portage/package.keywords
in either form, leaving it up to the user to decide if it is to be a
file or a directory. At the same time, certain Portage tools -
app-portage/tatt for instance - implicitly assume it to be the latter.
Simply create an empty directory in the right place.